Split agent and controller config
